# kapsule-incus-manager/api/schema/openapi.yaml
openapi: "3.1.0"
info:
  title: Kapsule-Incus-Manager API
  version: "0.1.0"
  description: >
    Canonical API definition for the KIM daemon.
    All three frontends (QML, web, CLI) are thin clients against this API.
    The D-Bus interface mirrors every operation defined here.
  license:
    name: GPL-3.0-or-later
servers:
  - url: http://127.0.0.1:8765
    description: Local KIM daemon
...
components:
  schemas:
...
    Operation:
      type: object
      required: [id, description, status, created_at]
      properties:
        id:          { type: string, format: uuid }
        description: { type: string }
        status:
          type: string
          enum: [pending, running, succeeded, failed, cancelled]
        created_at:  { type: string, format: date-time }
        updated_at:  { type: string, format: date-time }
        metadata:    { type: object, additionalProperties: true }
        error:       { type: string }
    InstanceStatus:
      type: string
      enum: [Running, Stopped, Frozen, Error, Unknown]
    InstanceType:
      type: string
      enum: [container, virtual-machine]
    InstanceSummary:
      type: object
      required: [name, type, status, project, remote]
      properties:
        name:       { type: string }
        type:       { $ref: "#/components/schemas/InstanceType" }
        status:     { $ref: "#/components/schemas/InstanceStatus" }
        image:      { type: string }
        project:    { type: string }
        remote:     { type: string }
        created_at: { type: string, format: date-time }
        cpu_usage:  { type: number, description: "CPU usage 0.0–1.0" }
        memory_usage_bytes: { type: integer }
        disk_usage_bytes:   { type: integer }
...
  responses:
...
    Accepted:
      description: Operation accepted and running asynchronously
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Operation" }
...
paths:
...
  /api/v1/provisioning/macos/image/firmware:
    post:
      operationId: downloadMacosFirmware
      summary: Download OVMF firmware and OpenCore via a helper container
      tags: [provisioning/macos]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                storage_pool: { type: string, default: "default" }
                project:      { type: string, default: "" }
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/macos/image/fetch:
    post:
      operationId: fetchMacosImage
      summary: Download macOS recovery image via a helper container
      tags: [provisioning/macos]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                version:      { type: string, default: "sonoma" }
                storage_pool: { type: string, default: "default" }
                volume_name:  { type: string }
                project:      { type: string, default: "" }
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/macos:
    post:
      operationId: createMacosVm
      summary: Create a macOS VM with storage volumes attached
      tags: [provisioning/macos]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:          { type: string }
                version:       { type: string, default: "sonoma" }
                ram:           { type: string, default: "4GiB" }
                cpus:          { type: integer, default: 4 }
                storage_pool:  { type: string, default: "default" }
                disk_vol:      { type: string }
                installer_vol: { type: string }
                opencore_vol:  { type: string }
                profiles:      { type: array, items: { type: string } }
                project:       { type: string, default: "" }
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/macos/{name}/start:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
      - { name: project, in: query, schema: { type: string, default: "" } }
    post:
      operationId: startMacosVm
      summary: Start a macOS VM
      tags: [provisioning/macos]
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/macos/{name}/stop:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
      - { name: force, in: query, schema: { type: boolean, default: false } }
      - { name: project, in: query, schema: { type: string, default: "" } }
    post:
      operationId: stopMacosVm
      summary: Stop a macOS VM
      tags: [provisioning/macos]
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/macos/{name}/snapshots:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
      - { name: project, in: query, schema: { type: string, default: "" } }
    get:
      operationId: listMacosSnapshots
      summary: List snapshots for a macOS VM
      tags: [provisioning/macos]
      responses:
        "200":
          description: Snapshot list
          content:
            application/json:
              schema: { type: array, items: { type: object, additionalProperties: true } }
    post:
      operationId: createMacosSnapshot
      summary: Create a snapshot
      tags: [provisioning/macos]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [snapshot]
              properties:
                snapshot: { type: string }
                stateful: { type: boolean, default: false }
                project:  { type: string, default: "" }
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/macos/{name}/snapshots/schedule:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
    put:
      operationId: setMacosSnapshotSchedule
      summary: Set automatic snapshot schedule for a macOS VM
      tags: [provisioning/macos]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [schedule]
              properties:
                schedule: { type: string }
                expiry:   { type: string }
                project:  { type: string, default: "" }
      responses:
        "200":
          description: Schedule updated
  /api/v1/provisioning/macos/{name}/snapshots/{snapshot}/restore:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
      - { name: snapshot, in: path, required: true, schema: { type: string } }
      - { name: project, in: query, schema: { type: string, default: "" } }
    post:
      operationId: restoreMacosSnapshot
      summary: Restore a macOS VM snapshot
      tags: [provisioning/macos]
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/macos/{name}/snapshots/{snapshot}:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
      - { name: snapshot, in: path, required: true, schema: { type: string } }
      - { name: project, in: query, schema: { type: string, default: "" } }
    delete:
      operationId: deleteMacosSnapshot
      summary: Delete a macOS VM snapshot
      tags: [provisioning/macos]
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/macos/{name}/backups:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
      - { name: project, in: query, schema: { type: string, default: "" } }
    get:
      operationId: listMacosBackups
      summary: List Incus backups for a macOS VM
      tags: [provisioning/macos]
      responses:
        "200":
          description: Backup list
          content:
            application/json:
              schema: { type: array, items: { type: object, additionalProperties: true } }
    post:
      operationId: backupMacosVm
      summary: Create an Incus backup of a macOS VM
      tags: [provisioning/macos]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                backup_name:       { type: string }
                instance_only:     { type: boolean, default: false }
                optimized_storage: { type: boolean, default: true }
                project:           { type: string, default: "" }
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/macos/{name}/restore:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
    post:
      operationId: restoreMacosVmBackup
      summary: Restore a macOS VM from an Incus backup
      tags: [provisioning/macos]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [backup_name]
              properties:
                backup_name: { type: string }
                project:     { type: string, default: "" }
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/macos/{name}/gpus:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
    post:
      operationId: attachMacosGpu
      summary: Attach a GPU to a macOS VM
      tags: [provisioning/macos]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                dev_name: { type: string, default: "gpu0" }
                gpu_type: { type: string, default: physical }
                pci:      { type: string }
                vendor:   { type: string }
                project:  { type: string, default: "" }
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/macos/{name}/gpus/{dev_name}:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
      - { name: dev_name, in: path, required: true, schema: { type: string } }
      - { name: project, in: query, schema: { type: string, default: "" } }
    delete:
      operationId: detachMacosGpu
      summary: Detach a GPU from a macOS VM
      tags: [provisioning/macos]
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/macos/{name}/forwards:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
    post:
      operationId: addMacosForward
      summary: Add a port-forward proxy device to a macOS VM
      tags: [provisioning/macos]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [host_port, guest_port]
              properties:
                host_port:   { type: integer }
                guest_port:  { type: integer }
                protocol:    { type: string, default: "tcp" }
                listen_addr: { type: string, default: "127.0.0.1" }
                project:     { type: string, default: "" }
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/macos/{name}/forwards/{dev_name}:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
      - { name: dev_name, in: path, required: true, schema: { type: string } }
      - { name: project, in: query, schema: { type: string, default: "" } }
    delete:
      operationId: removeMacosForward
      summary: Remove a port-forward proxy device from a macOS VM
      tags: [provisioning/macos]
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/macos/{name}/disk/resize:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
    post:
      operationId: resizeMacosDisk
      summary: Resize the macOS VM disk volume
      tags: [provisioning/macos]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [new_size]
              properties:
                new_size:     { type: string, example: "256G" }
                storage_pool: { type: string, default: "default" }
                volume_name:  { type: string }
                project:      { type: string, default: "" }
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/macos/fleet:
    get:
      operationId: listMacosFleet
      summary: List macOS VMs
      tags: [provisioning/macos]
      parameters:
        - { name: project, in: query, schema: { type: string, default: "" } }
        - { name: status, in: query, schema: { type: string } }
      responses:
        "200":
          description: VM list
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/InstanceSummary" } }
  /api/v1/provisioning/macos/fleet/start:
    post:
      operationId: fleetStartMacos
      summary: Start multiple macOS VMs
      tags: [provisioning/macos]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [names]
              properties:
                names:   { type: array, items: { type: string } }
                project: { type: string, default: "" }
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/macos/fleet/stop:
    post:
      operationId: fleetStopMacos
      summary: Stop multiple macOS VMs
      tags: [provisioning/macos]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [names]
              properties:
                names:   { type: array, items: { type: string } }
                project: { type: string, default: "" }
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/macos/publish:
    post:
      operationId: publishMacosVm
      summary: Publish a macOS VM as a reusable Incus image
      tags: [provisioning/macos]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:        { type: string }
                alias:       { type: string }
                description: { type: string }
                public:      { type: boolean, default: false }
                project:     { type: string, default: "" }
      responses:
        "202": { $ref: "#/components/responses/Accepted" }

  # ── Windows VM provisioning ───────────────────────────────────────────────
...